1) Reformat project to be sepperated by pane 2) Include CameraPane and PurchasePane as sepperated panes 3) include index.android.js
