+++ src/purchase/index.js
@@ -1,0 +1,12 @@
+
+import { Component } from 'react';
+
+export default class PurchasePane extends Component {
+  render() {
+    return (
+      <View>
+        <Text>PurchasePane</Text>
+      </View>
+    );
+  }
+}
